convert col component to typescript
